requirements.txt
click==8.0.3
Flask==2.0.2
flask-sock==0.4.0
h11==0.13.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.28
protobuf==3.19.4
python-dotenv==0.19.2
PyYAML==6.0
simple-websocket==0.5.0
Turbo-Flask==0.8.0
Werkzeug==2.0.3
wsproto==1.0.0